refactor: move filter summary to local-first policy
