+++ ui/application/local_first_progress_facts.py
@@ -6,7 +6,7 @@ from dataclasses import replace
 from ...activities.application import build_activity_preview_selection_state
 from .local_first_activity_controls import build_current_activity_preview_request
 from ...visualization.application import DEFAULT_TEMPORAL_MODE_LABEL
-from .wizard_filter_summary import build_wizard_filter_description
+from .local_first_filter_summary import build_local_first_filter_description
 from .wizard_progress import build_wizard_progress_facts_from_runtime_state
 
 logger = logging.getLogger(__name__)
@@ -180,7 +180,7 @@ def current_local_first_filter_facts(dock, runtime_state) -> tuple[bool, int | N
     return (
         True,
         selection_state.filtered_count,
-        build_wizard_filter_description(preview_request),
+        build_local_first_filter_description(preview_request),
     )
